+++ tests/test_analysis/test_analysis_csv.py
@@ -8,7 +8,6 @@ from asyncpg.exceptions import UndefinedTableError
 from yarl import URL
 
 from tests.conftest import RESOURCE_ID, RESOURCE_URL
-from udata_hydra import config
 from udata_hydra.analysis.csv import analyse_csv, csv_to_db
 from udata_hydra.db.resource import Resource
 
@@ -85,46 +84,6 @@ async def test_analyse_csv_big_file(setup_catalog, rmock, db, fake_check, produc
     assert profile["total_lines"] == expected_count
 
 
-async def test_exception_analysis(setup_catalog, rmock, db, fake_check, produce_mock):
-    """
-    Tests that exception resources (files that are too large to be normally processed) are indeed processed.
-    """
-    save_config = config.MAX_FILESIZE_ALLOWED
-    config.override(MAX_FILESIZE_ALLOWED={"csv": 5000})
-    await db.execute(
-        f"UPDATE catalog SET resource_id = '{config.LARGE_RESOURCES_EXCEPTIONS[0]}' WHERE id=1"
-    )
-    check = await fake_check(resource_id=config.LARGE_RESOURCES_EXCEPTIONS[0])
-    filename, expected_count = ("20190618-annuaire-diagnostiqueurs.csv", 45522)
-    url = check["url"]
-    table_name = hashlib.md5(url.encode("utf-8")).hexdigest()
-    with open(f"tests/data/{filename}", "rb") as f:
-        data = f.read()
-    rmock.get(url, status=200, body=data)
-
-    # Check resource status before analysis
-    resource = await Resource.get(config.LARGE_RESOURCES_EXCEPTIONS[0])
-    assert resource["status"] is None
-
-    # Analyse the CSV
-    await analyse_csv(check_id=check["id"])
-
-    # Check resource status after analysis
-    resource = await Resource.get(config.LARGE_RESOURCES_EXCEPTIONS[0])
-    assert resource["status"] is None
-
-    count = await db.fetchrow(f'SELECT count(*) AS count FROM "{table_name}"')
-    assert count["count"] == expected_count
-    profile = await db.fetchrow(
-        "SELECT csv_detective FROM tables_index WHERE resource_id = $1", check["resource_id"]
-    )
-    profile = json.loads(profile["csv_detective"])
-    for attr in ("header", "columns", "formats", "profile"):
-        assert profile[attr]
-    assert profile["total_lines"] == expected_count
-    config.override(MAX_FILESIZE_ALLOWED=save_config)
-
-
 @pytest.mark.parametrize(
     "line_expected",
     (
@@ -155,7 +114,7 @@ async def test_csv_to_db_simple_type_casting(db, line_expected, clean_db):
             "header": list(columns.keys()),
             "columns": columns,
         }
-        await csv_to_db(fp.name, inspection, "test_table")
+        await csv_to_db(file_path=fp.name, inspection=inspection, table_name="test_table")
     res = list(await db.fetch("SELECT * FROM test_table"))
     assert len(res) == 1
     cols = ["__id", "int", "float", "string", "bool"]
@@ -200,7 +159,7 @@ async def test_csv_to_db_complex_type_casting(db, line_expected, clean_db):
             "columns": columns,
         }
         # Insert the data
-        await csv_to_db(fp.name, inspection, "test_table")
+        await csv_to_db(file_path=fp.name, inspection=inspection, table_name="test_table")
     res = list(await db.fetch("SELECT * FROM test_table"))
     assert len(res) == 1
     cols = ["__id", "json", "date", "datetime"]
@@ -227,7 +186,7 @@ async def test_basic_sql_injection(db, clean_db):
             "columns": columns,
         }
         # Insert the data
-        await csv_to_db(fp.name, inspection, "test_table")
+        await csv_to_db(file_path=fp.name, inspection=inspection, table_name="test_table")
     res = await db.fetchrow("SELECT * FROM test_table")
     assert res[injection] == "test"
 
@@ -249,7 +208,7 @@ async def test_percentage_column(db, clean_db):
             "columns": columns,
         }
         # Insert the data
-        await csv_to_db(fp.name, inspection, "test_table")
+        await csv_to_db(file_path=fp.name, inspection=inspection, table_name="test_table")
     res = await db.fetchrow("SELECT * FROM test_table")
     assert res["% mon pourcent"] == "test"
 
@@ -271,7 +230,7 @@ async def test_reserved_column_name(db, clean_db):
             "columns": columns,
         }
         # Insert the data
-        await csv_to_db(fp.name, inspection, "test_table")
+        await csv_to_db(file_path=fp.name, inspection=inspection, table_name="test_table")
     res = await db.fetchrow("SELECT * FROM test_table")
     assert res["xmin__hydra_renamed"] == "test"